Renaming tile_map, timeline plugin folder to kebab-case
