+++ .i18nrc.json
@@ -46,7 +46,7 @@
       "src/plugins/telemetry",
       "src/plugins/telemetry_management_section"
     ],
-    "tileMap": "src/plugins/tile_map",
+    "tileMap": "src/plugins/tile-map",
     "timeline": ["src/plugins/timeline", "src/plugins/vis_type_timeline"],
     "uiActions": "src/plugins/ui_actions",
     "visDefaultEditor": "src/plugins/vis_default_editor",
