+++ src/plugins/tile-map/opensearch_dashboards.json
