+++ src/plugins/tile-map/public/_tile_map.scss
