+++ src/plugins/tile-map/public/geohash_layer.js
