+++ src/plugins/tile-map/public/index.scss
