+++ src/plugins/tile-map/public/markers/geohash_grid.js
