+++ src/plugins/tile-map/public/markers/heatmap.js
