+++ src/plugins/tile-map/public/markers/scaled_circles.js
