+++ src/plugins/tile-map/public/markers/shaded_circles.js
