+++ src/plugins/tile-map/public/tile_map_fn.js
