+++ src/plugins/tile-map/public/tile_map_type.js
