+++ src/plugins/tile-map/public/tile_map_visualization.js
