+++ src/plugins/tile-map/public/tooltip_formatter.js
