+++ src/plugins/timeline/public/directives/_index.scss
@@ -1,8 +1,8 @@
 @import './timeline_expression_input';
 @import './cells/index';
-@import './timeline_expression_suggestions/index';
-@import './timeline_help/index';
-@import './timeline_interval/index';
+@import './timeline-expression-suggestions/index';
+@import './timeline-help/index';
+@import './timeline-interval/index';
 @import './saved_object_finder';
 @import './form';
