+++ src/plugins/timeline/public/directives/timeline-expression-suggestions/_index.scss
