+++ src/plugins/timeline/public/directives/timeline-expression-suggestions/_timeline_expression_suggestions.scss
