+++ src/plugins/timeline/public/directives/timeline-expression-suggestions/timeline_expression_suggestions.js
