+++ src/plugins/timeline/public/directives/timeline-help/_index.scss
