+++ src/plugins/timeline/public/directives/timeline-help/_timeline_help.scss
