+++ src/plugins/timeline/public/directives/timeline-help/timeline_help.js
