+++ src/plugins/timeline/public/directives/timeline-interval/_index.scss
