+++ src/plugins/timeline/public/directives/timeline-interval/_timeline_interval.scss
