+++ src/plugins/timeline/public/directives/timeline-interval/timeline_interval.html
